Removing repeated input checking in Lasso and DictLearning
I use :
- a `check_input` flag in `coordinate_descent.enet_path`, that is set to False when called from `ElasticNet.fit`
 - a `check_input` flag in `coordinate_descent.ElasticNet.fit`, that is set to False when called from `sparse_encode`

I changed the condition for overriding provided Gram Matrix in `linear_model.base._pre_fit`, in order not to do a pass on data when fit_intercept and normalize are set to False (in master, we override Gram matrix if we find that X was changed by centering and rescaling, even when these are disabled, which wastes computation)

I also avoid computing `cov` in `sparse_encode` when using coordiante_descent, as `cov` computation is done within the `Lasso` class.

On provided  `plot_online_sparse_pca` example, we gain a factor 2 in performance.